testing/hang.hs: Have sockets time out before killThread is called, not after
On Windows, if killThread is called first, hGetLine will still encounter an IO
exception for some reason.  Thus, having the socket time out first is slightly
more desirable because it produces more consistent behavior.

There's no getting around the fact that you'll get an IO exception on timeout
on Windows, but ThreadKilled or <<timeout>> on Linux.  Maybe there is, but I
don't think it's worth spending a lot of time on.